+++ controlplane/controller/internal/controller/models.go
@@ -181,6 +181,11 @@ type Device struct {
 	MgmtVrf               string
 	IsisNet               string
 	DevicePathologies     []string
+	// Additional fields for metric labels
+	Code            string
+	ContributorCode string
+	ExchangeCode    string
+	LocationCode    string
 }
 
 func NewDevice(ip net.IP, publicKey string) *Device {